Implemented edit action and code cleanup.

+++ Frontend/src/components/task/index.js
@@ -4,27 +4,38 @@ import { useNavigate } from 'react-router-dom';
 import { logoutAction } from '../../redux/user/userAction';
 import DatePicker from "react-datepicker";
 import "react-datepicker/dist/react-datepicker.css";
-import { getTaskListAction, AddTaskAction, editTaskAction, deleteTaskAction } from '../../redux/task/taskAction';
-import taskReducer from '../../redux/task/taskReducer';
+import { getTaskListAction, AddTaskAction, getTaskAction, editTaskAction, deleteTaskAction } from '../../redux/task/taskAction';
 
 const TaskDashboard = (props) => {
-  
+  const {info} = props;
+  const taskInput = useRef(null);
   const datepickerRef = useRef(null);
   const navigate = useNavigate();
   const [startDate, setStartDate] = useState(new Date().toISOString().split('T')[0]);
   const [taskItem, setTaskItem] = useState('');
   const [list, setList] = useState([]);
 
+  const [editItem, setEditItem] = useState(0);
+
   useEffect(() => {
-    console.log(props.tasks)
     setList((props.tasks ? props.tasks : []));
   }, [props.tasks])
 
+  // Any update happend in the item
   useEffect(() => {
-    props.getTaskListAction();
-  }, [props?.updated])
+    if(info) {
+      taskInput.current.focus();
+      setTaskItem(info?.message);
+      setStartDate(fomateDate(info?.due));
+      setEditItem(info?.id);
+    } else {
+      props.getTaskListAction();
+    }
+  }, [props?.updated, info])
 
+  // Default component render
   useEffect(() => {
+    //setStartDate(fomateDate());
     props.getTaskListAction();
   }, [])
 
@@ -33,14 +44,39 @@ const TaskDashboard = (props) => {
     navigate('/login');
   }
 
-  const handleAddTask= () => {
-    props.AddTaskAction({"message": taskItem, "date": startDate});
-    setTaskItem('');
-    setStartDate(new Date().toISOString().split('T')[0]);
+  const handleSubmitTask= (e) => {
+    e.preventDefault();
+    if(editItem) {
+      props.editTaskAction({"id": editItem, "message": taskItem, "date": startDate});
+    } else {
+      props.AddTaskAction({"message": taskItem, "date": startDate});
+    }
+    handleReset();
   }
 
   const handleDelete = (id) => {
-    props.deleteTaskAction(id);
+    if(window.confirm("Are you sure want to delete this task?")) {
+      props.deleteTaskAction(id);
+    }
+  }
+
+  const handleEdit = (id) => {
+    props.getTaskAction(id);
+  }
+
+  const handleReset = () => {
+    setTaskItem('');
+    setStartDate(fomateDate());
+    setEditItem(0);
+  }
+
+
+  const fomateDate = (date) => {
+    const dateObj =  date ? new Date(date) : new Date();
+    var d = dateObj.getDate();
+    var m = dateObj.getMonth() + 1; //Month from 0 to 11
+    var y = dateObj.getFullYear();
+    return '' + (d <= 9 ? '0' + d : d) + '-' + (m<=9 ? '0' + m : m) + '-' + y;
   }
 
   return (
@@ -48,7 +84,7 @@ const TaskDashboard = (props) => {
     {/*<h1>Task Dashboard</h1>
     <button  className="btn btn-primary" onClick={(e) => logout()} >Logout</button>
     */}
-    <section className="vh-100">
+    <section className="vh-100 width100">
       <div className="container py-5 h-100">
         <div className="row d-flex justify-content-center align-items-center h-100">
           <div className="col">
@@ -66,25 +102,28 @@ const TaskDashboard = (props) => {
                   <div className="card">
                     <div className="card-body">
                       <div className="d-flex flex-row align-items-center">
-                        <input type="text" required className="form-control form-control-lg" placeholder="Add new..." onChange={(e) => setTaskItem(e.target.value)} value={taskItem} />
-                          <DatePicker ref={datepickerRef} 
-                          dateFormat="dd-MM-yyyy"
-                          selected={startDate} 
-                          onChange={(date) => setStartDate(`${String(date.getDate()).padStart(2, "0")}-${String(date.getMonth() + 1).padStart(2, "0")}-${date.getFullYear()}`)} />
-                          {/* <a href="#!" data-mdb-toggle="tooltip" title="Set due date">
-                            <i className="fa fa-calendar fa-lg me-3" onClick={() => datepickerRef.current.setFocus(true)}></i>
-                          </a> */ }
-                        <div>
-                          <button type="button" className="btn btn-primary" onClick={handleAddTask}>Add</button>
-                        </div>
+                        <form className='form-group width100' method='post' onSubmit={handleSubmitTask}>
+                          <input type="text" ref={taskInput} required className="form-control form-control-lg" placeholder="Add new task..." onChange={(e) => setTaskItem(e.target.value)} value={taskItem} />
+                          <div className='mt-3 row'>
+                            <span className='col-md-3'>
+                            <DatePicker ref={datepickerRef} 
+                              className="form-control form-control-sm mr-3" 
+                              dateFormat="dd-MM-yyyy"
+                              selected={startDate} 
+                              onChange={(date) =>setStartDate(fomateDate(date))} />
+                            </span>
+                            <button type="submit" className="btn btn-primary mx-3 col-md-2">{editItem ? "Update Task" : "Add Task"}</button>
+                            {editItem ? <button type='reset' className="btn btn-primary mx-3 col-md-2" onClick={() => handleReset() }>Reset</button> : ''}
+                          </div>
+                        </form>
                       </div>
                     </div>
                   </div>
                 </div>
 
                 <hr className="my-4" />
-
-                  {props?.msg ? <div className='alert alert-danger'>{props?.msg}</div> : ''}
+                  {props?.msg ? <div className='alert alert-success'>{props?.msg}</div> : ''}
+                  {props?.msg && props?.errorFlag ? <div className='alert alert-danger'>{props?.msg}</div> : ''}
                   
                   {list && list.map((item) => {
                     return(<ul className="list-group list-group-horizontal rounded-0" id={item.id}>
@@ -112,7 +151,7 @@ const TaskDashboard = (props) => {
                     </li>*/}
                     <li className="list-group-item ps-3 pe-0 py-1 rounded-0 border-0 bg-transparent">
                       <div className="d-flex flex-row justify-content-end mb-1">
-                        <span className="text-info" data-mdb-toggle="tooltip" title="Edit todo"><i className="fa fa-pencil me-3"></i></span>
+                        <span className="text-info" data-mdb-toggle="tooltip" title="Edit todo" onClick={() => handleEdit(item.id)}><i className="fa fa-pencil me-3"></i></span>
                         <span className="text-danger" data-mdb-toggle="tooltip" title="Delete todo" onClick={() => handleDelete(item.id)}><i className="fa fa-trash"></i></span>
                       </div>
                       <div className="text-end text-muted">
@@ -137,14 +176,17 @@ const mapStateToProps = state => {
   return { 
     tasks: state?.taskReducer?.tasks,
     updated: state?.taskReducer?.updated,
-    msg: state?.taskReducer?.msg
+    msg: state?.taskReducer?.msg,
+    info: state?.taskReducer?.info,
+    errorFlag: state?.taskReducer?.errorFlag,
   }
 }
 const mapDispatchToProps = {
   logoutAction,
   getTaskListAction,
   AddTaskAction,
-  editTaskAction,
-  deleteTaskAction
+  getTaskAction,
+  deleteTaskAction,
+  editTaskAction
 }
 export default connect(mapStateToProps, mapDispatchToProps)(TaskDashboard);
